[ADDED] AllowedConnectionTypes in UserClaims
This would be used by the server to restrict the connections
from the kind of clients that are found in the map, if the map
is not empty.
In other words, an empty map means no restriction (which also
allows backward compatibility).
If the map is not empty, then when a client connects, if the
user is found, the server will check that the type of client
(say standard, or websocket) is found in this map. If not, the
connection will be rejected.

I am not set on either the name or the type. It could be an
array of string and I will convert to a map[string]struct{}
in the server when processing the JWT.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>