[ADDED] IsClaimRevoked to check user claim and fixed some comments.
Comment for RevokeAt(), IsRevokedAt() have been fixed.

The function IsRevoked() must no longer be used (and removed in v2).
It will now always return true. Use IsRevokedAt() instead with
proper timestamp.

Note: This was done in v2 and backported to branch_v1 but was not
added to the v1 in the main branch. So this is a port of #108.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>