overhaul jwt v2 and v1 inter-tangling
Move all the top-level go code into v2/v1compat and explicitly refer to that,
and with all other go.??? files removed, so that from the go module handling
perspective, this is all in one Go module.  This will prevent v1 code
references from going out to Go modules and fetching old vulnerable code.

We would use v2/internal/v1compat but we want this compatibility code to be
visible to tests in other software, such as nats-io/nats-account-server too.

This PR changes modules, imports, and CI test invocations, but makes no changes
to the Go functionality of the library.